requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto==2.48.0
celery==3.1.25
celery-with-mongodb==3.0
click==6.7
deap==1.3.0
docutils==0.14
Flask==0.12.2
Flask-Cors==3.0.3
itsdangerous==0.24
Jinja2==2.10
Keras==2.3.0
kombu==3.0.37
linecache2==1.0.0
matplotlib==3.1.1
MarkupSafe==1.0
memory-profiler==0.52.0
nose==1.3.7
numpy==1.17.2
pandas==0.25.1
pika==0.12.0
Pint==0.8.1
plotly==4.1.1
ply==3.10
psutil==5.4.5
PyJWT==1.6.4
pymongo==3.6.1
pytest==5.2.1
python-dateutil==2.8.0
pytz==2018.3
rectpack==0.2.1
redis==2.10.6
scikit-learn==0.21.3
scipy==1.0.0
six==1.11.0
tensorflow==2.0.0
Theano==1.0.4
traceback2==1.4.0
unittest2==1.1.0
vine==1.1.4
Werkzeug==0.14.1
